+++ src/pages/Governance/Proposal/VoteButtons.tsx
@@ -1,41 +1,13 @@
 import {Button, Grid, Tooltip} from "@mui/material";
 import React from "react";
-import {AptosClient, AptosAccount, HexString, Types} from "aptos";
-import {useGlobalState} from "../../../GlobalState";
-import {doTransaction} from "../../utils";
-
-// TODO: replace the following hard code data with account info from the wallet connected
-const TEST_ACCOUNT_ADDRESS =
-  "c1bc62cb0e142a8fcfdcd01ce2a9f5b01355fba73290768f62069fa6902e1585";
-const TEST_ACCOUNT_SECRET_KEY =
-  "0x894e620e2e96748118f448388ac4877bb3799be5233426e869a97e2ffcd91381";
-
-async function vote(
-  networkValue: string,
-  proposalId: string,
-  shouldPass: boolean,
-) {
-  const client = new AptosClient(networkValue);
-  const address = HexString.ensure(TEST_ACCOUNT_ADDRESS);
-  const secretKey = HexString.ensure(TEST_ACCOUNT_SECRET_KEY);
-  const account = new AptosAccount(secretKey.toUint8Array(), address);
-
-  const payload: Types.TransactionPayload = {
-    type: "script_function_payload",
-    function: "0x1::aptos_governance::vote",
-    type_arguments: [],
-    arguments: [account.address().hex(), proposalId, shouldPass],
-  };
-
-  await doTransaction(account, client, payload);
-}
+import useSubmitVote from "../SubmitVote";
 
 type Props = {
   proposalId: string;
 };
 
 export function VoteButtons({proposalId}: Props) {
-  const [state, _] = useGlobalState();
+  const [submitVote] = useSubmitVote(proposalId);
 
   const voted = false; // TODO - fetch real data
 
@@ -44,15 +16,7 @@ export function VoteButtons({proposalId}: Props) {
   };
 
   const onVote = (shouldPass: boolean) => {
-    vote(state.network_value, proposalId, shouldPass);
-  };
-
-  const onForVoteClick = () => {
-    onVote(true);
-  };
-
-  const onAgainstVoteClick = () => {
-    onVote(false);
+    submitVote(shouldPass);
   };
 
   if (voted) {
@@ -70,10 +34,10 @@ export function VoteButtons({proposalId}: Props) {
       justifyContent={{xs: "center", sm: "flex-end"}}
       alignItems="center"
     >
-      <Button variant="primary" sx={{mr: 2}} onClick={onForVoteClick}>
+      <Button variant="primary" sx={{mr: 2}} onClick={() => onVote(true)}>
         For
       </Button>
-      <Button variant="primary" onClick={onAgainstVoteClick}>
+      <Button variant="primary" onClick={() => onVote(false)}>
         Against
       </Button>
     </Grid>
